Utility for converting index files to ARFF file.
The first line of index file is an integer denoting number of features
in dataset and the second line in index file indicates the number of
classes in the dataset.

There are two important assumptions about the dataset:
    * Feature count is fixed over the dataset --- no unknown values.
    * The class symbols are integers drawn from [1, class_count].

In data section, each feature in the index data is seperated by whitespaces
and the last token denotes the class variable.

So, for the following index file:

        # Feature Count
        2
        # Class count
        2

        # Data
        A B 1
        B C 2

The ARFF format would be:

        %
        @RELATION wsd
        @ATTRIBUTE f1 NUMERIC
        @ATTRIBUTE f2 NUMERIC
        @ATTRIBUTE class {1, 2}

        @DATA
        1, 2, 1
        2, 3, 2

Where the WordMap has mapped A, B, and C to 1, 2, and 3.